+++ component/backend/Controller/Key.php
@@ -1,0 +1,17 @@
+<?php
+/**
+ * @package    contactus
+ * @copyright  Copyright (c)2013-2020 Nicholas K. Dionysopoulos / Akeeba Ltd
+ * @license    GNU General Public License version 3 or later
+ */
+
+namespace Akeeba\ContactUs\Admin\Controller;
+
+defined('_JEXEC') or die;
+
+use FOF30\Controller\DataController;
+
+class Key extends DataController
+{
+
+}