+++ component/backend/Model/Keys.php
@@ -1,0 +1,33 @@
+<?php
+/**
+ * @package    contactus
+ * @copyright  Copyright (c)2013-2020 Nicholas K. Dionysopoulos / Akeeba Ltd
+ * @license    GNU General Public License version 3 or later
+ */
+
+namespace Akeeba\ContactUs\Admin\Model;
+
+defined('_JEXEC') or die;
+
+use FOF30\Model\DataModel;
+
+/**
+ * Model Akeeba\ContactUs\Admin\Model\Categories
+ *
+ * Fields:
+ *
+ * @property  int    $contactus_key_id
+ * @property  string $email
+ * @property  string $key
+ *
+ * Filters:
+ *
+ * @method  $this  contactus_key_id()  contactus_key_id(int $v)
+ * @method  $this  email()             email(string $v)
+ * @method  $this  key()               key(string $v)
+ *
+ **/
+class Keys extends DataModel
+{
+
+}