+++ component/backend/View/Keys/Html.php
@@ -1,0 +1,13 @@
+<?php
+/**
+ * @package    contactus
+ * @copyright  Copyright (c)2013-2020 Nicholas K. Dionysopoulos / Akeeba Ltd
+ * @license    GNU General Public License version 3 or later
+ */
+
+namespace Akeeba\ContactUs\Admin\View\Keys;
+
+class Html extends \FOF30\View\DataView\Html
+{
+
+}