+++ component/frontend/Model/Items.php
@@ -10,24 +10,31 @@ namespace Akeeba\ContactUs\Site\Model;
 defined('_JEXEC') or die();
 
 use Akeeba\ContactUs\Site\Helper\Akismet;
+use Exception;
 use FOF30\Model\DataModel;
 use FOF30\Model\Mixin\Assertions;
+use gnupg;
 use Joomla\CMS\Application\SiteApplication;
 use Joomla\CMS\Captcha\Captcha;
 use Joomla\CMS\Factory;
+use Joomla\CMS\Language\Text;
+use OpenPGP;
+use OpenPGP_Crypt_Symmetric;
+use OpenPGP_LiteralDataPacket;
+use OpenPGP_Message;
 
 /**
  * Model Akeeba\ContactUs\Admin\Model\Items
  *
  * Fields:
  *
- * @property  int     $contactus_item_id
- * @property  int     $contactus_category_id
- * @property  string  $fromname
- * @property  string  $fromemail
- * @property  string  $subject
- * @property  string  $body
- * @property  string  $token
+ * @property  int        $contactus_item_id
+ * @property  int        $contactus_category_id
+ * @property  string     $fromname
+ * @property  string     $fromemail
+ * @property  string     $subject
+ * @property  string     $body
+ * @property  string     $token
  *
  * Filters:
  *
@@ -48,7 +55,7 @@ use Joomla\CMS\Factory;
  *
  * Relations:
  *
- * @property  Categories  $category
+ * @property  Categories $category
  *
  **/
 class Items extends DataModel
@@ -79,7 +86,7 @@ class Items extends DataModel
 			/** @var SiteApplication $app */
 			$app = Factory::getApplication();
 		}
-		catch (\Exception $e)
+		catch (Exception $e)
 		{
 			return null;
 		}
@@ -91,7 +98,45 @@ class Items extends DataModel
 			return null;
 		}
 
-		return Captcha::getInstance($plugin, array('namespace' => $namespace));
+		return Captcha::getInstance($plugin, ['namespace' => $namespace]);
+	}
+
+	/**
+	 * Returns the IDs of categories with encrypted keys support
+	 *
+	 * @param   bool  $strict  True to only return categories where ALL recipients support encrypted emails.
+	 *
+	 * @return  array
+	 */
+	public function getEncryptedCategories(bool $strict = true): array
+	{
+		/** @var Keys $keysModel */
+		$keysModel      = $this->container->factory->model('Keys')->tmpInstance();
+		$emailsWithKeys = $keysModel->enabled(1)->get(true)->fetch('email')->toArray();
+
+		if (empty($emailsWithKeys))
+		{
+			return [];
+		}
+
+		/** @var Categories $catModel */
+		$catModel      = $this->container->factory->model('Categories')->tmpInstance();
+		$allCategories = $catModel->enabled(1)
+			->get(true)->filter(function (Categories $cat) use ($emailsWithKeys, $strict) {
+				$emails = explode(',', $cat->email);
+				$emails = array_map('trim', $emails);
+
+				$commonEmails = array_intersect($emails, $emailsWithKeys);
+
+				if ($strict)
+				{
+					return count($commonEmails) === count($emails);
+				}
+
+				return count($commonEmails) > 0;
+			});
+
+		return $allCategories->fetch($catModel->getKeyName())->toArray();
 	}
 
 
@@ -106,7 +151,7 @@ class Items extends DataModel
 		$apiKey = $this->container->params->get('akismet_api_key', '');
 
 		$this->saveSuccessful = true;
-		$this->isSpam = Akismet::isSpamContent($apiKey, $this->fromname, $this->fromemail, $this->body);
+		$this->isSpam         = Akismet::isSpamContent($apiKey, $this->fromname, $this->fromemail, $this->body);
 
 		// Don't email the admins if it's spam.
 		if (!$this->isSpam)
@@ -159,31 +204,60 @@ class Items extends DataModel
 		/** @var DataModel $category */
 		$category = $this->container->factory->model('Categories')->tmpInstance();
 		$category->findOrFail($this->contactus_category_id);
+
 		$emails = explode(',', $category->email);
+		$emails = array_map('trim', $emails);
 
 		if (empty($emails))
 		{
-			return false;
-		}
-
-		foreach ($emails as $email)
-		{
-			$mailer->addRecipient($email);
+			return;
 		}
 
 		// Set the subject
-		$subject = \JText::sprintf('COM_CONTACTUS_ITEMS_MSG_EMAIL_SUBJECT',
+		$subject = Text::sprintf('COM_CONTACTUS_ITEMS_MSG_EMAIL_SUBJECT',
 			\JFactory::getConfig()->get('sitename'),
 			$category->title,
 			$this->subject);
 
 		$mailer->setSubject($subject);
 
-		// Set the body
-		$mailer->msgHTML($this->body);
+		// On fully unencrypted categories we simply send out emails with a CC to all recipients
+		$encryptedCats = $this->getEncryptedCategories(false);
 
-		// Send the email
-		$mailer->Send();
+		if (!in_array($category->getId(), $encryptedCats))
+		{
+			foreach ($emails as $email)
+			{
+				$mailer->addRecipient($email);
+			}
+
+			// Set the body
+			$mailer->msgHTML($this->body);
+
+			// Send the email
+			$mailer->Send();
+
+			return;
+		}
+
+		foreach ($emails as $email)
+		{
+			$thisMailer = clone $mailer;
+			$thisMailer->addRecipient($email);
+
+			try
+			{
+				$encrypted = $this->encryptBody($this->body, $email);
+				$thisMailer->setBody('This is a PGP encrypted message');
+				$thisMailer->addStringAttachment($encrypted, 'encrypted.asc', '8bit', 'application/pgp-encrypted', 'attachment');
+			}
+			catch (Exception $e)
+			{
+				$thisMailer->msgHTML($this->body);
+			}
+
+			$thisMailer->Send();
+		}
 	}
 
 	/**
@@ -215,7 +289,7 @@ class Items extends DataModel
 		$mailer->addRecipient($this->fromemail);
 
 		// Set the subject
-		$subject = \JText::sprintf('COM_CONTACTUS_ITEMS_MSG_AUTOREPLY_SUBJECT', \JFactory::getConfig()->get('sitename'));
+		$subject = Text::sprintf('COM_CONTACTUS_ITEMS_MSG_AUTOREPLY_SUBJECT', \JFactory::getConfig()->get('sitename'));
 
 		$mailer->setSubject($subject);
 
@@ -229,17 +303,17 @@ class Items extends DataModel
 	/**
 	 * Pre-processes the text of the automatic reply, replacing variables in it.
 	 *
-	 * @param   string    $text     The original text
-	 * @param   DataModel $category The contact category of the received contact message
+	 * @param   string     $text      The original text
+	 * @param   DataModel  $category  The contact category of the received contact message
 	 *
 	 * @return  string  The processed message
 	 */
 	private function _preProcessAutoreply($text, DataModel $category)
 	{
-		$replacements = array(
+		$replacements = [
 			'[SITENAME]' => \JFactory::getConfig()->get('sitename'),
 			'[CATEGORY]' => $category->title,
-		);
+		];
 
 		$rawData = $this->getData();
 
@@ -254,5 +328,37 @@ class Items extends DataModel
 		}
 
 		return $text;
+	}
+
+	/**
+	 * Encrypts the message body with the GunPG key
+	 *
+	 * @param   string  $body   The string to encrypt.
+	 * @param   string  $email  The email recipient. They must have a corresponding Key entry.
+	 *
+	 * @return  string  The encrypted message.
+	 *
+	 * @throws  Exception  Thrown if encryption is not possible.
+	 */
+	private function encryptBody(string $body, string $email): string
+	{
+		/** @var Keys $keysModel */
+		$keysModel = $this->container->factory->model('Keys')->tmpInstance();
+		$keyRecord = $keysModel->enabled(1)->email($email)->firstOrFail();
+
+		// Key unarmor and parsing
+		preg_match('/-----BEGIN ([A-Za-z ]+)-----/', $keyRecord->pubkey, $matches);
+		$marker     = (empty($matches[1])) ? 'MESSAGE' : $matches[1];
+		$key        = OpenPGP::unarmor($keyRecord->pubkey, $marker);
+		$openpgpKey = OpenPGP_Message::parse($key);
+
+		// Encrypt into a PGP message
+		$plain_data = new OpenPGP_LiteralDataPacket($body, [
+			'format' => 'u', 'filename' => tempnam(sys_get_temp_dir(), 'cspgp'),
+		]);
+
+		$encrypted = OpenPGP_Crypt_Symmetric::encrypt($openpgpKey, new OpenPGP_Message([$plain_data]));
+
+		return OpenPGP::enarmor($encrypted->to_bytes(), 'PGP MESSAGE', []);
 	}
 }