+++ component/frontend/Model/Keys.php
@@ -1,0 +1,17 @@
+<?php
+/**
+ * @package    contactus
+ * @copyright  Copyright (c)2013-2020 Nicholas K. Dionysopoulos / Akeeba Ltd
+ * @license    GNU General Public License version 3 or later
+ */
+
+namespace Akeeba\ContactUs\Site\Model;
+
+defined('_JEXEC') or die;
+
+use FOF30\Model\DataModel;
+
+class Keys extends DataModel
+{
+
+}